Split current implementation into separate files
